Advanced-AI-Powered-X-Ray-Defect-Detection/ai_model/requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
opencv-python==4.8.0.76
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
